+++ packages/kit/src/utils/filesystem.js
@@ -169,11 +169,12 @@ export function from_fs(str) {
 export function resolve_entry(entry) {
 	if (fs.existsSync(entry)) {
 		const stats = fs.statSync(entry);
-		const index = path.join(entry, 'index');
-
 		if (stats.isFile()) {
 			return entry;
-		} else if (fs.existsSync(index)) {
+		}
+
+		const index = path.join(entry, 'index');
+		if (fs.existsSync(index + '.js') || fs.existsSync(index + '.ts')) {
 			return resolve_entry(index);
 		}
 	}