+++ packages/kit/src/utils/url.js
@@ -199,26 +199,3 @@ function allow_nodejs_console_log(url) {
 		};
 	}
 }
-
-const DATA_SUFFIX = '/__data.json';
-const HTML_DATA_SUFFIX = '.html__data.json';
-
-/** @param {string} pathname */
-export function has_data_suffix(pathname) {
-	return pathname.endsWith(DATA_SUFFIX) || pathname.endsWith(HTML_DATA_SUFFIX);
-}
-
-/** @param {string} pathname */
-export function add_data_suffix(pathname) {
-	if (pathname.endsWith('.html')) return pathname.replace(/\.html$/, HTML_DATA_SUFFIX);
-	return pathname.replace(/\/$/, '') + DATA_SUFFIX;
-}
-
-/** @param {string} pathname */
-export function strip_data_suffix(pathname) {
-	if (pathname.endsWith(HTML_DATA_SUFFIX)) {
-		return pathname.slice(0, -HTML_DATA_SUFFIX.length) + '.html';
-	}
-
-	return pathname.slice(0, -DATA_SUFFIX.length);
-}