+++ packages/kit/test/apps/basics/test/client.test.js
@@ -228,13 +228,14 @@ test.describe('Load', () => {
 		await clicknav('[href="/load/fetch-cache-control"]');
 
 		// 3. Come back to the original page (client side)
+		/** @type {string[]} */
 		const requests = [];
-		page.on('request', (request) => requests.push(request));
+		page.on('request', (request) => requests.push(request.url()));
 		await clicknav('[href="/load/fetch-cache-control/headers-diff"]');
 
-		// 4. We expect the same data and no new request because it was cached.
+		// 4. We expect the same data and no new request (except a navigation request in case of server-side route resolution) because it was cached.
 		expect(await page.textContent('h2')).toBe('a / b');
-		expect(requests).toEqual([]);
+		expect(requests.filter((r) => !r.includes('/__route.js'))).toEqual([]);
 	});
 
 	test('permits 3rd party patching of fetch in universal load functions', async ({ page }) => {
@@ -826,32 +827,56 @@ test.describe('Invalidation', () => {
 test.describe('data-sveltekit attributes', () => {
 	test('data-sveltekit-preload-code', async ({ page }) => {
 		/** @type {string[]} */
-		const requests = [];
-		page.on('request', (r) => {
-			requests.push(r.url());
+		const responses = [];
+
+		const nodes_location = process.env.DEV
+			? '.svelte-kit/generated/client/nodes/'
+			: '/_app/immutable/nodes/';
+
+		page.on('response', async (response) => {
+			const url = response.url();
+			if (url.includes(nodes_location)) {
+				responses.push(url);
+			}
 		});
 
 		// eager
 		await page.goto('/data-sveltekit/preload-code');
-		expect(requests.length).toBeGreaterThanOrEqual(1);
+		await page.locator('#eager').hover();
+		await page.locator('#eager').dispatchEvent('touchstart');
+		// expect 4 nodes on initial load: root layout, root error, current page, and eager preload
+		expect(responses.length).toEqual(4);
 
 		// viewport
-		requests.length = 0;
+		responses.length = 0;
 		page.locator('#viewport').scrollIntoViewIfNeeded();
-		await Promise.all([page.waitForTimeout(100), page.waitForLoadState('networkidle')]);
-		expect(requests.length).toBeGreaterThanOrEqual(1);
+		await page.locator('#viewport').hover();
+		await page.locator('#viewport').dispatchEvent('touchstart');
+		await Promise.all([
+			page.waitForTimeout(100), // wait for preloading to start
+			page.waitForLoadState('networkidle') // wait for preloading to finish
+		]);
+		expect(responses.length).toEqual(1);
 
 		// hover
-		requests.length = 0;
-		await page.locator('#hover').dispatchEvent('mousemove');
-		await Promise.all([page.waitForTimeout(100), page.waitForLoadState('networkidle')]);
-		expect(requests.length).toBeGreaterThanOrEqual(1);
+		responses.length = 0;
+		await page.locator('#hover').hover();
+		await page.locator('#hover').dispatchEvent('touchstart');
+		await Promise.all([
+			page.waitForTimeout(100), // wait for preloading to start
+			page.waitForLoadState('networkidle') // wait for preloading to finish
+		]);
+		expect(responses.length).toEqual(1);
 
 		// tap
-		requests.length = 0;
+		responses.length = 0;
+		await page.locator('#tap').hover();
 		await page.locator('#tap').dispatchEvent('touchstart');
-		await Promise.all([page.waitForTimeout(100), page.waitForLoadState('networkidle')]);
-		expect(requests.length).toBeGreaterThanOrEqual(1);
+		await Promise.all([
+			page.waitForTimeout(100), // wait for preloading to start
+			page.waitForLoadState('networkidle') // wait for preloading to finish
+		]);
+		expect(responses.length).toEqual(1);
 	});
 
 	test('data-sveltekit-preload-data', async ({ page }) => {
@@ -874,7 +899,7 @@ test.describe('data-sveltekit attributes', () => {
 		});
 
 		await page.goto('/data-sveltekit/preload-data');
-		await page.locator('#one').dispatchEvent('mousemove');
+		await page.locator('#one').hover();
 		await Promise.all([
 			page.waitForTimeout(100), // wait for preloading to start
 			page.waitForLoadState('networkidle') // wait for preloading to finish
@@ -883,7 +908,7 @@ test.describe('data-sveltekit attributes', () => {
 
 		requests.length = 0;
 		await page.goto('/data-sveltekit/preload-data');
-		await page.locator('#two').dispatchEvent('mousemove');
+		await page.locator('#two').hover();
 		await Promise.all([
 			page.waitForTimeout(100), // wait for preloading to start
 			page.waitForLoadState('networkidle') // wait for preloading to finish
@@ -892,12 +917,22 @@ test.describe('data-sveltekit attributes', () => {
 
 		requests.length = 0;
 		await page.goto('/data-sveltekit/preload-data');
-		await page.locator('#three').dispatchEvent('mousemove');
+		await page.locator('#three').hover();
 		await Promise.all([
 			page.waitForTimeout(100), // wait for preloading to start
 			page.waitForLoadState('networkidle') // wait for preloading to finish
 		]);
 		expect(requests.length).toBe(0);
+
+		requests.length = 0;
+		await page.goto('/data-sveltekit/preload-data');
+		await page.locator('#tap').hover();
+		await page.locator('#tap').dispatchEvent('touchstart');
+		await Promise.all([
+			page.waitForTimeout(100), // wait for preloading to start
+			page.waitForLoadState('networkidle') // wait for preloading to finish
+		]);
+		expect(requests.length).toBe(1);
 	});
 
 	test('data-sveltekit-preload-data network failure does not trigger navigation', async ({